+++ src/u16x16_.rs
@@ -244,7 +244,9 @@ macro_rules! impl_shl_t_for_u16x16 {
       fn shl(self, rhs: $shift_type) -> Self::Output {
         pick! {
           if #[cfg(target_feature="avx2")] {
-            let shift = cast([rhs as u64, 0]);
+            // Use `rhs % 16` to perform wrapping shift and not unbounded shift.
+            #[expect(clippy::suspicious_arithmetic_impl)]
+            let shift = cast([rhs as u64 & 15, 0]);
             Self { avx2: shl_all_u16_m256i(self.avx2, shift) }
           } else {
             Self {
@@ -268,7 +270,9 @@ macro_rules! impl_shr_t_for_u16x16 {
       fn shr(self, rhs: $shift_type) -> Self::Output {
         pick! {
           if #[cfg(target_feature="avx2")] {
-            let shift = cast([rhs as u64, 0]);
+            // Use `rhs % 16` to perform wrapping shift and not unbounded shift.
+            #[expect(clippy::suspicious_arithmetic_impl)]
+            let shift = cast([rhs as u64 & 15, 0]);
             Self { avx2: shr_all_u16_m256i(self.avx2, shift) }
           } else {
             Self {