+++ tests/simd_float.rs
@@ -479,40 +479,55 @@ fn test_round() {
     for value in simd_chunks!([
       0.0,
       0.1,
+      (0.5 as T).next_down(),
       0.5,
       0.7,
-      -0.0,
-      -0.1,
-      -0.5,
-      -0.7,
-      2.0,
-      2.1,
-      2.5,
-      2.7,
-      -2.0,
-      -2.1,
-      -2.5,
-      -2.7,
-      5.0,
-      5.1,
-      5.5,
-      5.7,
-      -5.0,
-      -5.1,
-      -5.5,
-      -5.7,
+      8388607.0,
+      4503599627370495.0,
       T::MAX,
-      T::MIN,
-      T::NAN,
       T::INFINITY,
-      T::NEG_INFINITY,
+      T::NAN
     ])
+    .flat_map(|x| {
+      [x, x.map(|x| x + 1.0), x.map(|x| x + 2.0), x.map(|x| x + 3.0)]
+    })
+    .flat_map(|x| [x, x.map(|x| -x)])
     .chain(random_iter())
     {
-      // TODO:  Currently `round` actually behaves like `round_ties_even`.
-      // Decide the correct behavior then add documentation.
-      let expected = Simd::new(value.map(T::round_ties_even));
+      let expected = Simd::new(value.map(T::round));
       let actual = Simd::new(value).round();
+
+      assert!(
+        actual ^ expected == Simd::ZERO,
+        "expected: {expected:?}\n  actual: {actual:?}\n   value: {value:?}",
+      );
+    }
+  });
+}
+
+#[test]
+fn test_round_ties_even() {
+  for_simd_types!(|T: Float, N| {
+    for value in simd_chunks!([
+      0.0,
+      0.1,
+      (0.5 as T).next_down(),
+      0.5,
+      0.7,
+      8388607.0,
+      4503599627370495.0,
+      T::MAX,
+      T::INFINITY,
+      T::NAN
+    ])
+    .flat_map(|x| {
+      [x, x.map(|x| x + 1.0), x.map(|x| x + 2.0), x.map(|x| x + 3.0)]
+    })
+    .flat_map(|x| [x, x.map(|x| -x)])
+    .chain(random_iter())
+    {
+      let expected = Simd::new(value.map(T::round_ties_even));
+      let actual = Simd::new(value).round_ties_even();
 
       assert!(
         actual ^ expected == Simd::ZERO,